Rename samples to ServiceBus.Minimal.Library
