--- test/Datadog.Trace.ClrProfiler.IntegrationTests/SmokeTests/RebusSqlServerRequestReply.cs
@@ -1,27 +1,0 @@
-// <copyright file="RebusSqlServerRequestReply.cs" company="Datadog">
-// Unless explicitly stated otherwise all files in this repository are licensed under the Apache 2 License.
-// This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2017 Datadog, Inc.
-// </copyright>
-
-#if !NET452
-using Xunit;
-using Xunit.Abstractions;
-
-namespace Datadog.Trace.ClrProfiler.IntegrationTests.SmokeTests
-{
-    public class RebusSqlServerRequestReply : SmokeTestBase
-    {
-        public RebusSqlServerRequestReply(ITestOutputHelper output)
-            : base(output, "Rebus.SqlServer.RequestReply", maxTestRunSeconds: 90)
-        {
-        }
-
-        [Fact]
-        [Trait("Category", "Smoke")]
-        public void NoExceptions()
-        {
-            CheckForSmoke(shouldDeserializeTraces: false);
-        }
-    }
-}
-#endif
