--- test/test-applications/regression/MassTransit.Autofac.Saga/Components/SubmitOrderConsumerDefinition.cs
@@ -1,21 +1,0 @@
-using GreenPipes;
-using MassTransit;
-using MassTransit.ConsumeConfigurators;
-using MassTransit.Definition;
-using MassTransit.Autofac.Saga.Contracts;
-
-namespace MassTransit.Autofac.Saga.Components
-{
-    public class SubmitOrderConsumerDefinition :
-        ConsumerDefinition<SubmitOrderConsumer>
-    {
-        protected override void ConfigureConsumer(IReceiveEndpointConfigurator endpointConfigurator,
-            IConsumerConfigurator<SubmitOrderConsumer> consumerConfigurator)
-        {
-            endpointConfigurator.UseInMemoryOutbox();
-            endpointConfigurator.UseMessageRetry(r => r.Interval(3, 1000));
-
-            EndpointConvention.Map<SubmitOrder>(endpointConfigurator.InputAddress);
-        }
-    }
-}
