--- test/test-applications/regression/MassTransit.Autofac.Saga/Contracts/OrderReceived.cs
@@ -1,10 +1,0 @@
-using System;
-
-namespace MassTransit.Autofac.Saga.Contracts
-{
-    public interface OrderReceived
-    {
-        Guid OrderId { get; }
-        DateTimeOffset OrderDateTime { get; }
-    }
-}
