--- test/test-applications/regression/MassTransit.Autofac.Saga/Contracts/OrderStateCreated.cs
@@ -1,10 +1,0 @@
-using System;
-
-namespace MassTransit.Autofac.Saga.Contracts
-{
-    public interface OrderStateCreated
-    {
-        Guid OrderId { get; }
-        DateTime Timestamp { get; }
-    }
-}
