--- test/test-applications/regression/MassTransit.Autofac.Saga/Contracts/SubmitOrder.cs
@@ -1,10 +1,0 @@
-﻿using System;
-
-namespace MassTransit.Autofac.Saga.Contracts
-{
-    public interface SubmitOrder
-    {
-        Guid OrderId { get; }
-        DateTimeOffset OrderDateTime { get; }
-    }
-}
