--- test/test-applications/regression/NServiceBus.SqlServer.Saga/Server/CompleteOrder.cs
@@ -1,7 +1,0 @@
-namespace NServiceBus.SqlServer.Saga.Server
-{
-    public class CompleteOrder
-    {
-        public string OrderDescription { get; set; }
-    }
-}
