--- test/test-applications/regression/NServiceBus.SqlServer.Saga/Shared/OrderCompleted.cs
@@ -1,10 +1,0 @@
-using System;
-using NServiceBus;
-
-namespace NServiceBus.SqlServer.Saga.Shared
-{
-    public class OrderCompleted : IEvent
-    {
-        public Guid OrderId { get; set; }
-    }
-}
