--- test/test-applications/regression/NServiceBus.SqlServer.Saga/Shared/StartOrder.cs
@@ -1,10 +1,0 @@
-using System;
-using NServiceBus;
-
-namespace NServiceBus.SqlServer.Saga.Shared
-{
-    public class StartOrder : IMessage
-    {
-        public Guid OrderId { get; set; }
-    }
-}
