--- test/test-applications/regression/Rebus.SqlServer.RequestReply/Shared/Job.cs
@@ -1,12 +1,0 @@
-namespace Rebus.SqlServer.RequestReply.Shared
-{
-    public class Job
-    {
-        public int Id { get; private set; }
-
-        public Job(int id)
-        {
-            Id = id;
-        }
-    }
-}
