--- test/test-applications/regression/Rebus.SqlServer.RequestReply/Shared/Reply.cs
@@ -1,14 +1,0 @@
-namespace Rebus.SqlServer.RequestReply.Shared
-{
-    public class Reply
-    {
-        public int Id { get; private set; }
-        public int OsProcessId { get; private set; }
-
-        public Reply(int id, int osProcessId)
-        {
-            Id = id;
-            OsProcessId = osProcessId;
-        }
-    }
-}
