+++ test/test-applications/regression/ServiceBus.Minimal.MassTransit/Contracts/OrderReceived.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace ServiceBus.Minimal.MassTransit.Contracts
+{
+    public interface OrderReceived
+    {
+        Guid OrderId { get; }
+        DateTimeOffset OrderDateTime { get; }
+    }
+}
