+++ test/test-applications/regression/ServiceBus.Minimal.MassTransit/Contracts/OrderStateCreated.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace ServiceBus.Minimal.MassTransit.Contracts
+{
+    public interface OrderStateCreated
+    {
+        Guid OrderId { get; }
+        DateTime Timestamp { get; }
+    }
+}
