+++ test/test-applications/regression/ServiceBus.Minimal.MassTransit/ServiceBus.Minimal.MassTransit.csproj
