+++ test/test-applications/regression/ServiceBus.Minimal.NServiceBus/Server/CompleteOrder.cs
@@ -1,0 +1,7 @@
+namespace ServiceBus.Minimal.NServiceBus.Server
+{
+    public class CompleteOrder
+    {
+        public string OrderDescription { get; set; }
+    }
+}
