+++ test/test-applications/regression/ServiceBus.Minimal.NServiceBus/ServiceBus.Minimal.NServiceBus.csproj
