+++ test/test-applications/regression/ServiceBus.Minimal.NServiceBus/Shared/OrderCompleted.cs
@@ -1,0 +1,10 @@
+using System;
+using NServiceBus;
+
+namespace ServiceBus.Minimal.NServiceBus.Shared
+{
+    public class OrderCompleted : IEvent
+    {
+        public Guid OrderId { get; set; }
+    }
+}
