+++ test/test-applications/regression/ServiceBus.Minimal.NServiceBus/Shared/StartOrder.cs
@@ -1,0 +1,10 @@
+using System;
+using NServiceBus;
+
+namespace ServiceBus.Minimal.NServiceBus.Shared
+{
+    public class StartOrder : IMessage
+    {
+        public Guid OrderId { get; set; }
+    }
+}
