+++ test/test-applications/regression/ServiceBus.Minimal.Rebus/ServiceBus.Minimal.Rebus.csproj
