+++ test/test-applications/regression/ServiceBus.Minimal.Rebus/Shared/Job.cs
@@ -1,0 +1,12 @@
+﻿namespace ServiceBus.Minimal.Rebus.Shared
+{
+    public class Job
+    {
+        public int Id { get; private set; }
+
+        public Job(int id)
+        {
+            Id = id;
+        }
+    }
+}
